- Modifié la logique du SaveChanges pour mettre à jour le empty entity status au niveau du DataContext, en utilisant la nouvelle propriété IsEntityEmpty. - Adapté TileContainerController pour ne plus planter si aucune action n'est attachée au bouton (-) d'une collection.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010@14515 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Cresus.Core/Controllers/TileContainerController.cs
@@ -463,12 +463,15 @@ namespace Epsitec.Cresus.Core.Controllers
 					this.GenerateTiles ();
 				};
 
-				tile.RemoveClicked += sender =>
+				if (item.DeleteItem != null)
 				{
-					this.controller.Orchestrator.CloseSubViews (this.controller);
-					item.DeleteItem ();
-					this.GenerateTiles ();
-				};
+					tile.RemoveClicked += sender =>
+					{
+						this.controller.Orchestrator.CloseSubViews (this.controller);
+						item.DeleteItem ();
+						this.GenerateTiles ();
+					};
+				}
 
 				tile.EnableAddRemoveButtons = item.DataType == SummaryDataType.CollectionItem && item.AutoGroup;
 				item.SummaryTile = tile;
